Fix the CI test case: testBatchMessageThreadpoolSize() - by removing the ASSERT for non-invariant.(#2531)
TaskExecutor is internally uses _batchMessageExecutorService when batch mode is enabled. This threadpool is of type newCachedThreadPool().

newCachedThreadPool - Creates a thread pool that creates new threads as needed, but will reuse previously constructed threads when they are available.

The test creates 10 database resources and the check is to see if it pool size is 10 or more. This is not invariant. It will depend.

Removing the assert as it is not invariant.

Co-authored-by: Komal Desai <kdesai@kdesai-mn1.linkedin.biz>