WAGED - Part 2 - Fix the core calculation for n - n+1 issue. This change is focused on changes to core logic and has not done any additional code cleanup, de-dup etc.
There are 3 more changes planned for the change once core logic is checked-in.
Change-1:  code dedup across computeBestPossibleStateForPartition
Change-2: moving the message procesing state to 'DataProcess' stage of the pipeline.
Change-3: This is may be: to add metrics to indicate that we diverted from 'preferred list'

The core logic is as follows:
Pre-Compute stage
------------------
- in WAGED rebalancer: computeNewIdealState, we create 'WagedInstanceCapacity' which tracks instance capacity and WeightProvider for partitions. These changes were introduced in Part-1.
- We process the pending messages for the resources where if we are moving out partition but not yet moved, we reduce the capacity.

Once we compute the ideal state, we go through preference list sorting.
There are no changes to the pure computational change.

Post-compute stage
------------------
We look at the current-state as well as ideal state, we combine the list and do the sorting of the nodes. We then pick up only the required numReplica nodes.

This is where we look at the nodes which are "NEW" (ie. not in currentState) and check if it has the required capacity to hold the new partition.
If not, we remove it from combined list.

Testing done:
Have run through the mvn test (10 times)
In another change, I have added a new test case and verified that it works.
[ERROR] Failures:
[ERROR]   TestNoThrottleDisabledPartitions.testDisablingTopStateReplicaByDisablingInstance:98 expected:<false> but was:<true>
[ERROR]   TestClusterMaintenanceMode.testMaintenanceHistory:412 expected:<EXIT> but was:<ENTER>
[ERROR]   TestParticipantFreeze.testUnfreezeParticipant:228 expected:<true> but was:<false>
[ERROR]   TestRecurringJobQueue.testDeletingRecurrentQueueWithHistory:298 expected:<true> but was:<false>
[ERROR] Tests run: 1335, Failures: 4, Errors: 0, Skipped: 0