Open Add NPR guard when reading instance config - Race when reading config while adding/removing instance (#2669)
When adding/removing instance from a cluster, it is possible that at a point of time, instance config is gone but the INSTANCE ZNode is still there. This would cause config map in helix controller cache to have instance mapped to null config.
The failure is not a blocking error causing further pipeline failure but we would like to avoid noise in log.
This change add null check before access instance config.