[apache/helix] -- Handling Throwable during Helix message processing (#2730)
We are improving the Helix Task processing to allow Helix to continue working on the failed Task in case of Throwables of type Error, and not only for type Exceptions. We are extending the scope from Exception to Throwable, to allow ZK state clean-up and retry code path to be executed on encountering java.lang.Error of any type as well.