[apache/helix] -- Add SetPartitionToError for participants to self annotate a node to ERROR state (#2792)
Co-authored-by: Charanya Sudharsanan <csudhars@csudhars-mn2.linkedin.biz>

What: An API endpoint that validates the incoming request and sends a state transition message to sets one or more partitions from any current state to ERROR state.

Why: Currently, the participants are unable to set a partition to an ERROR state explicitly when they seem to be stuck in a specific current state. The only way a replica can be set to ERROR is from within a state model. Having an endpoint to allow this behavior would allow the clients to call the resetPartition endpoint to set it back to INIT state and recover the replica. resetPartition works only on partitions in error state.