[apache/helix] -- Enable JDK 1.8 (backward) compatibility for dependent modules of meta-client and helix-core (#2806)
We would like to provide a backward compatible support to our consumers where they also have an option to use JDK-8 compiled meta-client jar, if they have such a requirement. By default we will generate JDK-11 jars and JDK-8 jars using a classifier. In this PR, we are enabling the JDK* build of full sub-tree of the helix-core and meta-client modules. This includes:
helix-common, zookeeper-api, metadata-store-directory-common, metrics-common