skip min active check if instance replica in unhealthy state (#3019)
Stoppable API will skip the min active replica check for a replica if that replica is already in an unhealthy state. This is because stopping the instance will not affect that partition's availability as the replica is already in an unhealthy state (offline/error). Previously, this would prevent a node with an error state replica from being stopped if min_active for that partition was not satisfied. With this change, we will not consider the sister replicas and drop anyways.