Only build property key path on construction (#3040)
Changes PropertyKey so that it only builds the path once when it is instantiated. PropertyKey.getPath() previously would rebuild the path on every call. This could cause a performance hit as it was performing a regex match each time. getPath now returns the path that was constructed when the PropertyKey object is first instantiated