Add default get value for task partition map to prevent NPE(#3078)
This PR fixes an NPE in task pipeline caused by a difference in map keyset. In fillActiveTaskCount(..) we iterate over the keyset of one map, but then call .get(key) on a 2nd map. This works on the assumption that the two maps will have the same keyset. The map we get the keys from is built from nodes with task current states and the other is from live and enabled nodes. If there is a node with a current state but has been disabled, then there will be key mismatch and a null value will be returned, which we then attempt to perform arithmetic on (+ operator) and get an NPE. Interestingly, this can only occur after a controller reset as it needs the 2nd map keys to be cleared and repopulated without the disabled node. 
This fix adds a default get value of 0 to both get operations to prevent any NPEs